removed RKWeeklyViewController
