+++ Sources/RKManager/RKViewController.swift
@@ -13,11 +13,10 @@ public struct RKViewController: View {
     @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
     
     @Binding var isPresented: Bool
-    
     @ObservedObject var rkManager: RKManager
-    
+    @State var pages = [RKWeeklyPage]()
     @State var index: Int = 0
-
+    
     
     public var body: some View {
         Group {
@@ -28,10 +27,55 @@ public struct RKViewController: View {
                     Spacer()
                 }.padding(15)
             }
+            rkManager.isWeeklyView ? AnyView(weeklyBody) : AnyView(monthlyBody)
+        }
+    }
+    
+    public var monthlyBody: some View {
+        Group {
             self.rkManager.isVertical
                 ? self.rkManager.isContinuous ? AnyView(self.verticalView) : AnyView(self.verticalViewPage)
                 : self.rkManager.isContinuous ? AnyView(self.horizontalView) : AnyView(self.horizontalViewPage)
             Spacer()
+        }
+    }
+    
+    public var weeklyBody: some View {
+        Group {
+            self.rkManager.isContinuous
+                ? AnyView(continuousView)
+                : AnyView(RKPageView(rkManager: rkManager, pages: pages))
+        }.onAppear(perform: loadWeeklyData)
+    }
+    
+    var continuousView: some View {
+        ScrollView (.horizontal) {
+            HStack {
+                ForEach(0..<self.numberOfMonths()) { index in
+                    VStack (spacing: 15) {
+                        Divider()
+                        HStack {
+                            ForEach(0..<self.numberOfWeeks(monthOffset: index)) { _ in
+                                VStack (spacing: 15) {
+                                    RKWeekdayHeader(rkManager: self.rkManager)
+                                    RKMonthHeader(rkManager: self.rkManager, monthOffset: index)
+                                }
+                            }
+                        }
+                        RKMonth(isPresented: self.$isPresented, rkManager: self.rkManager, monthOffset: index)
+                        Spacer()
+                    }
+                    Divider()
+                }
+            }
+        }
+    }
+    
+    func loadWeeklyData() {
+        for i in 0..<self.numberOfMonths() {
+            for j in 0..<self.numberOfWeeks(monthOffset: i) {
+                pages.append(RKWeeklyPage(isPresented: $isPresented, rkManager: rkManager, monthNdx: i, weekNdx: j))
+            }
         }
     }
     
@@ -77,22 +121,34 @@ public struct RKViewController: View {
     // vertical page scroll
     var verticalViewPage: some View {
         RKPageView(rkManager: rkManager,
-            pages: (0..<numberOfMonths()).map {
-                index in RKPage(isPresented: $isPresented, rkManager: rkManager, index: index)
+                   pages: (0..<numberOfMonths()).map {
+                    index in RKPage(isPresented: $isPresented, rkManager: rkManager, index: index)
         })
     }
     
     // horizontal page scroll
     var horizontalViewPage: some View {
         RKPageView(rkManager: rkManager,
-            pages: (0..<numberOfMonths()).map {
-                index in RKPage(isPresented: $isPresented, rkManager: rkManager, index: index)
+                   pages: (0..<numberOfMonths()).map {
+                    index in RKPage(isPresented: $isPresented, rkManager: rkManager, index: index)
         })
     }
     
     func onDone() {
         // to go back to the previous view
         self.presentationMode.wrappedValue.dismiss()
+    }
+    
+    func numberOfWeeks(monthOffset: Int) -> Int {
+        let firstOfMonth = firstOfMonthForOffset(monthOffset: monthOffset)
+        let rangeOfWeeks = rkManager.calendar.range(of: .weekOfMonth, in: .month, for: firstOfMonth)
+        return (rangeOfWeeks?.count)!
+    }
+    
+    func firstOfMonthForOffset(monthOffset : Int) -> Date {
+        var offset = DateComponents()
+        offset.month = monthOffset
+        return rkManager.calendar.date(byAdding: offset, to: rkManager.RKFirstDateMonth())!
     }
     
     func numberOfMonths() -> Int {
@@ -105,7 +161,7 @@ public struct RKViewController: View {
         components.day = 0
         return rkManager.calendar.date(from: components)!
     }
- 
+    
 }
 
 #if DEBUG
